+++ monolithium/monolithium/utils.rs
@@ -1,7 +1,7 @@
 use crate::*;
 
 // Round an integer to a nearest multiple of another
-pub fn nearest(num: i64, mul: i64) -> i64 {
+pub fn nearest(num: i32, mul: i32) -> i32 {
     (num + mul/2) / mul * mul
 }
